requirements.txt
flask
spacy
httpx
python-dotenv
gTTS